- Reviewed the structure for chargePoint.
